Add `workspace_tests` crate.
