+++ crate/model/src/common/edge.rs
@@ -10,22 +10,3 @@ pub enum Edge {
     /// This edge is purely for adjusting how `dot` lays out the graph.
     Layout,
 }
-
-#[cfg(test)]
-mod tests {
-    use super::Edge;
-
-    #[test]
-    fn clone() {
-        let edge = Edge::Logic;
-
-        assert_eq!(Edge::Logic, edge.clone());
-    }
-
-    #[test]
-    fn debug() {
-        let edge = Edge::Logic;
-
-        assert_eq!("Logic", format!("{edge:?}"));
-    }
-}
