+++ workspace_tests/src/lib.rs
@@ -1,0 +1,4 @@
+#![cfg_attr(coverage_nightly, feature(coverage_attribute))]
+#![cfg(test)]
+
+mod model;
