+++ workspace_tests/src/model.rs
@@ -1,0 +1,1 @@
+mod edge;
